ArctosR: An Interface to the Arctos Database

The ArctosR package provides a set of functions to help users perform requests to the Arctos API to download data. It provides a set of builder classes for performing complex requests, as well as a set of simple functions for automating many common requests and workflows.

About Arctos

Arctos is a collection management information system serving over 5 million records from natural and cultural history collections. Arctos integrates access to collections from disciplines such as anthropology, botany, entomology, ethnology, herpetology, geology, ichthyology, mammalogy, mineralogy, ornithology, paleontology, parasitology, as well as archival and cultural collections. The Arctos database is accessible through a web interface at https://arctos.database.museum/ More information about Arctos, can be found at https://arctosdb.org/about/, and in Cicero et al. (2024) doi:10.1371/journal.pone.0296478.

Functions in ArctosR

get_query_parameters, get_result_parameters, get_record_count, get_records, check_for_status, get_error_response, get_last_response_url, response_data, save_response_rds, read_response_rds, save_response_csv, expand_column

See also

Author

Maintainer: Harlan R. Williams harlanrhwilliams@gmail.com

Authors:

Marlon E. Cobos manubio13@gmail.com

Jocelyn P. Colella colella@ku.edu

Michelle S. Koo mkoo@berkeley.edu

Vijay Barve vijay.barve@gmail.com